Develop

+++ NanoChallenge#3/NanoChallenge#3/Controllers/ViewController.swift
@@ -10,11 +10,20 @@ import UIKit
 //Setando a array de todas classes coreData como global para ajudar no fluxo de desenvolvimento
 var idSubject : [IdSubjects] = IdSubjects().getAllIdSubjects()
 var subjects : [Subjects] = Subjects().getAllSubjects()
+var classes : [Classes] = Classes().getAllClasses()
 
 class ViewController: UIViewController {
 
-
+    //IBOutlets
+    @IBOutlet weak var iptNewSubject: UITextField!
+    @IBOutlet weak var vwButton: UIView!
+    
+    @IBOutlet weak var vwIptBtn: UIView!
     @IBOutlet weak var tbViewSubjects: UITableView!
+    
+    @IBOutlet weak var background: UIImageView!
+    
+    @IBOutlet weak var tituloLg: UILabel!
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -25,68 +34,81 @@ class ViewController: UIViewController {
         //Nesse caso seram criadas extensões da própria classe com as UIViews necessárias
         tbViewSubjects.delegate = self
         tbViewSubjects.dataSource = self
+
+        //Setando borda para o input juntamente com o botao de adicionar
+        vwIptBtn.clipsToBounds = true
+        vwIptBtn.layer.cornerRadius = 10
+        
+        vwButton.clipsToBounds = true
+        vwButton.layer.cornerRadius = 10
+        vwButton.layer.maskedCorners = [.layerMaxXMaxYCorner, .layerMaxXMinYCorner]
+        
+        //Adicionando imagem de background para a view principal
+        background.translatesAutoresizingMaskIntoConstraints = false
+        background.topAnchor.constraint(equalTo: view.topAnchor).isActive = true
+        background.leadingAnchor.constraint(equalTo: view.leadingAnchor).isActive = true
+        background.trailingAnchor.constraint(equalTo: view.trailingAnchor).isActive = true
+        background.bottomAnchor.constraint(equalTo: view.bottomAnchor).isActive = true
+        
+        //Deixando a font do titulo com estilo bold
+        tituloLg.font = UIFont.boldSystemFont(ofSize: 34)
     }
+    
 
-
-    //Função que redireciona para a página de adicionar uma nova matéria
-    @IBAction func adicionarNovaMateria(_ sender: Any) {
+    
+    @IBAction func adicionarMatéria(_ sender: Any) {
+    
+        guard let txt = iptNewSubject.text else { return }
         
-        let entry = storyboard?.instantiateViewController(withIdentifier: "CadMatViewController") as! CadMatViewController
+        let newTxt = txt.trimmingCharacters(in: .whitespacesAndNewlines)
         
-        //Condição para checar se existe algum objeto criado dentro da entidade IdSubject
-        if idSubject.isEmpty {
-            
-            //Criando um novo id para a adicionar na entidade
-            let newId : Int64 = 1
-            
-            //Adicionando id na entidade IdSubjects
-            IdSubjects().createIdSubject(id: newId)
-            
-            //Setando o valor do id à variavel
-            entry.id = newId
-        } else {
-            
-            //Criando uma nova id com o valor do id atual do IdSubjects + 1 (incrementação)
-            let newId = idSubject[0].id + 1;
-            
-            print(newId)
-            
-            //Setando o valor do id à variavel
-            entry.id = newId
-            
-            //Deletando o id com o valor passadp
-            IdSubjects().deleteAllIdsSubjects()
-            
-            idSubject = IdSubjects().getAllIdSubjects()
-            
+        if(newTxt != "") {
+            //Condição para checar se existe algum objeto criado dentro da entidade IdSubject
             if idSubject.isEmpty {
-                print("Deu certo");
-            }
+
+                //Criando um novo id para a adicionar na entidade
+                let newId : Int64 = 1
+
+                //Adicionando id na entidade IdSubjects
+                IdSubjects().createIdSubject(id: newId)
+                
+                //Chamando a função para criar uma nova matéria
+                Subjects().createSubject(name: txt, id: newId)
+
+            } else {
+                //Criando uma nova id com o valor do id atual do IdSubjects + 1 (incrementação)
+                let newId = idSubject[0].id + 1;
+
+               //Deletando o id com o valor passadp
+               IdSubjects().deleteAllIdsSubjects()
+
+               idSubject = IdSubjects().getAllIdSubjects()
+
+               if idSubject.isEmpty {
+                   print("Deu certo");
+               }
+
+               //Inserindo um novo id com o valor incrementado
+               IdSubjects().createIdSubject(id: newId)
+
+               idSubject = IdSubjects().getAllIdSubjects()
+
+               //Chamando a função para criar uma nova matéria
+               Subjects().createSubject(name: newTxt, id: newId)
+
+
+           }
             
-            //Inserindo um novo id com o valor incrementado
-            IdSubjects().createIdSubject(id: newId)
-            
-            idSubject = IdSubjects().getAllIdSubjects()
-            
-            print(idSubject[0].id)
-        
-     
-        }
-        
-        //Função que atualiza a tela assim que os dados das matérias forem cadastrados
-        entry.update = {
+            //Utilizando função para atualizar tabela de matérias
             subjects = Subjects().getAllSubjects()
             DispatchQueue.main.async {
                 self.tbViewSubjects.reloadData()
             }
+            
+            iptNewSubject.text = ""
         }
         
-        //Redirecionamento sendo efetuado
-        navigationController?.pushViewController(entry, animated: true)
-        
-        
     }
-    
     
 }
 
@@ -94,20 +116,61 @@ extension ViewController: UITableViewDelegate {
     
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         print(subjects[indexPath.row].idSubject)
+        
+        //Criando a variável que popula a ViewController selecionada
+        let entry = storyboard?.instantiateViewController(withIdentifier: "ClassesViewController") as! ClassesViewController
+        
+        //Populando a ViewController com base nas variáveis opcionais criadas na mesma
+        entry.nameSubject = subjects[indexPath.row].name
+        entry.idSubject = subjects[indexPath.row].idSubject
+        
+        //Enviando o usuário para a ViewController selecionada junto com as variáveis setadas
+        navigationController?.pushViewController(entry, animated: true)
+        
+        //Tirando o status de selecionado da linha
+        tbViewSubjects.deselectRow(at: indexPath, animated: true)
     }
     
-    func tableView(_ tableView: UITableView, editingStyleForRowAt indexPath: IndexPath) -> UITableViewCell.EditingStyle {
-        return .delete
-    }
-    
-    func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCell.EditingStyle, forRowAt indexPath: IndexPath) {
-        if editingStyle == .delete{
-            tbViewSubjects.beginUpdates()
+    func tableView(_ tableView: UITableView, editActionsForRowAt indexPath: IndexPath) -> [UITableViewRowAction]? {
+        let delete = UITableViewRowAction(style: .destructive, title: "Deletar") { action, indexPath in
+            //Começando as alterações dentro da tabela
+            self.tbViewSubjects.beginUpdates()
+            
+            //Deletando item do core data
             Subjects().deleteSubject(item: subjects[indexPath.row])
-            tbViewSubjects.deleteRows(at: [indexPath], with: .fade)
+            
+            //Deletando item da tabela
+            self.tbViewSubjects.deleteRows(at: [indexPath], with: .fade)
+            
+            //Buscando os itens novamente no core data
             subjects = Subjects().getAllSubjects()
-            tbViewSubjects.endUpdates()
+            
+            self.tbViewSubjects.endUpdates()
         }
+        
+        let edit = UITableViewRowAction(style: .normal, title: "Editar") { action, indexPath in
+            //Criando a variável que popula a ViewController selecionada
+            let entry = self.storyboard?.instantiateViewController(withIdentifier: "EditSubjectViewController") as! EditSubjectViewController
+            
+            //Populando as variáveis da nova viewController
+            entry.disciplina = subjects[indexPath.row]
+            
+            //Criando função que atualiza os dados quando volta para a viewController principal
+            entry.update = {
+                subjects = Subjects().getAllSubjects()
+                DispatchQueue.main.async {
+                    self.tbViewSubjects.reloadData()
+                }
+            }
+            
+            //Determinando o tamanho que a tela vai aparecer para o usuário
+            entry.sheetPresentationController?.detents = [.medium()]
+            
+            //Apresentando a nova viewController na viewController principal
+            self.present(entry, animated: true, completion: nil)
+        }
+        edit.backgroundColor = .link
+        return [delete, edit]
     }
     
 }
@@ -117,17 +180,8 @@ extension ViewController: UITableViewDelegate {
 //Na extensão UITableViewDataSource é obrigatório duas funções: a função que contem cellForRowAt e numbersOfRowsInSection
 extension ViewController: UITableViewDataSource {
     
-    //Função que cria um titulo para uma seção específica de células
     func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        switch section {
-            case 0:
-                return "Matérias atuais"
-            default:
-                return ""
-        }
-            
-        
-        
+        return "Disciplinas adicionadas"
     }
     
     public func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
